backend/requirements.txt
alembic==1.13.1
asgiref==3.7.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.0
Flask-Migrate==4.0.5
flask-redis==0.4.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.3
marshmallow==3.21.0
marshmallow-sqlalchemy==1.0.0
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
redis==5.0.1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
waitress==3.0.0
Werkzeug==3.0.1
